Add separate step WriteCsv
